+++ src/Sidebar/Navigation/index.jsx
@@ -1,8 +1,9 @@
 import React from 'react';
 import styled from 'styled-components';
+import { NavLink } from 'react-router-dom';
 
 import media from '../../media';
-import MenuItem from './MenuItem';
+import Menu from './Menu';
 
 const Nav = styled.nav`
   display: none;
@@ -16,12 +17,54 @@ const Nav = styled.nav`
   `}
 `;
 
+const Link = styled(NavLink)`
+  color: #3c3c3c;
+  display: flex;
+  justify-content: center;
+  font-family: 'AndaleMono';
+  padding: 11px 0;
+  text-decoration: none;
+  text-transform: uppercase;
+
+  &:last-child {
+    padding-bottom: 34px;
+  }
+
+  &:hover {
+    color: #fff;
+  }
+`;
+
 export default () => (
   <Nav>
-    <MenuItem to="#" hasSubmenu>Football</MenuItem>
-    <MenuItem to="#" hasSubmenu>Running</MenuItem>
-    <MenuItem to="#" hasSubmenu>Basketball</MenuItem>
-    <MenuItem to="#">Brands</MenuItem>
-    <MenuItem to="#">Micoach</MenuItem>
+    <Menu title="football" hasSubmenu>
+      <Link to="/football/shoes">shoes</Link>
+      <Link to="/football/clothing">
+        clothing
+      </Link>
+      <Link to="/football/accesories">
+        accesories
+      </Link>
+    </Menu>
+    <Menu title="running" hasSubmenu>
+      <Link to="/running/shoes">shoes</Link>
+      <Link to="/running/clothing">
+        clothing
+      </Link>
+      <Link to="/running/accesories">
+        accesories
+      </Link>
+    </Menu>
+    <Menu title="basketball" hasSubmenu>
+      <Link to="/basketball/shoes">shoes</Link>
+      <Link to="/basketball/clothing">
+        clothing
+      </Link>
+      <Link to="/basketball/accesories">
+        accesories
+      </Link>
+    </Menu>
+    <Menu title="brands" />
+    <Menu title="micoach" />
   </Nav>
 );
